refactor: Move CTxMemPoolEntry definitions to kernel/mempool_entry.h
Extracts CTxMemPoolEntry, LockPoints, and CompareIteratorByHash from
txmempool.h and moves them to a new kernel/mempool_entry.h file.
Updates all files that use these classes to include the new header.

This refactoring aligns with Bitcoin Core's modularization efforts
and prepares for future kernel library work.

Backports bitcoin/bitcoin#26609
