+++ src/test/util/setup_common.cpp
@@ -16,6 +16,7 @@
 #include <index/txindex.h>
 #include <init.h>
 #include <interfaces/chain.h>
+#include <kernel/mempool_entry.h>
 #include <net.h>
 #include <net_processing.h>
 #include <noui.h>
